Add default headers to request
This change exposes the RequestBuilder to the main module. Thus it is
now the responsibillity of main to send the request. This gives us a
more flexible way to build the request. The intention is to have helpers
for building requests in the request module, and have the main module
call the appropriate helpers to build the request.

The default headers are added to clarify how the request building is
supposed to work.