Add item arguments
This adds a var args argument. I could not get clap to accept that the
method had a default value. As soon as I added extra argumets after url
I had to provide the method to prevent the url being parsed as the
method (wich failed, of course). For now the method has been made
required.

The new item argument is still unused.